css and product page updates
Updated the CSS so that the site shrinks a bit more gracefully. Tuned it
to fit down to a vertical pad aspect ratio. needs testing.

recut all the homepage banners and selected them.
selected some product on the homepage for testing.

started more work on the search results for the product detail page.
appears zoovy:prod_ugenre is in use on his current layout, so using that
for now. having it added to catalog. Once done, should work. carousel
feature will need tuning.